requirements.txt
flask==2.3.3
flask-cors==4.0.0
pywebview==4.4.1
requests==2.31.0
markdown==3.5.1
python-markdown-math==0.8
pygments==2.16.1